requirements.txt
python=3.6.10
pytorch=1.1.0
torchvision=0.3.0
numpy=1.18.4
tqdm=4.46.0
scikit-learn=0.21.3
inflect=4.1.0
nltk=3.4.5
h5py=2.10.0
openjdk
pycocotools=2.0.0
scikit-image=0.16.2
gensim=3.8.0